requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
transformers
torch

openai
langchain
python-dotenv
